clean up environment module
